Hidden/given field changes for dimension line labels. See https://github.com/phetsims/area-model-common/issues/41

+++ js/game/model/AreaChallenge.js
@@ -403,14 +403,29 @@ define( function( require ) {
         display.partitionValuesProperties.get( Orientation.VERTICAL ).value = this.verticalPartitionSizeProperties;
       }
 
-      this.horizontalTotalListener = this.totalProperties.get( Orientation.HORIZONTAL ).linkAttribute( display.totalProperties.get( Orientation.HORIZONTAL ), 'value' );
-      this.verticalTotalListener = this.totalProperties.get( Orientation.VERTICAL ).linkAttribute( display.totalProperties.get( Orientation.VERTICAL ), 'value' );
+      // TODO: simplify
+      if ( this.description.horizontalTotalValue !== GameValue.HIDDEN ) {
+        this.horizontalTotalListener = this.totalProperties.get( Orientation.HORIZONTAL ).linkAttribute( display.totalProperties.get( Orientation.HORIZONTAL ), 'value' );
+      }
+      else {
+        display.totalProperties.get( Orientation.HORIZONTAL ).value = null;
+      }
+      if ( this.description.verticalTotalValue !== GameValue.HIDDEN ) {
+        this.verticalTotalListener = this.totalProperties.get( Orientation.VERTICAL ).linkAttribute( display.totalProperties.get( Orientation.VERTICAL ), 'value' );
+      }
+      else {
+        display.totalProperties.get( Orientation.VERTICAL ).value = null;
+      }
     },
 
     // TODO
     detachDisplay: function( display ) {
-      this.totalProperties.get( Orientation.HORIZONTAL ).unlink( this.horizontalTotalListener );
-      this.totalProperties.get( Orientation.VERTICAL ).unlink( this.verticalTotalListener );
+      if ( this.description.horizontalTotalValue !== GameValue.HIDDEN ) {
+        this.totalProperties.get( Orientation.HORIZONTAL ).unlink( this.horizontalTotalListener );
+      }
+      if ( this.description.verticalTotalValue !== GameValue.HIDDEN ) {
+        this.totalProperties.get( Orientation.VERTICAL ).unlink( this.verticalTotalListener );
+      }
     }
   }, {
